requirements.txt
alembic==1.4.2
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
email-validator==1.1.0
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.3.1
